06_training/requirements.txt
# Epoch Training Module - Dependencies
# Version: 1.0.0

# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.9

# Visualization
plotly>=5.18.0

# API
requests>=2.31.0

# Timezone handling
pytz>=2023.3

# Optional: For future image upload
# pillow>=10.0.0
